Text.Gigaparsec.Errors.TokenExtractors

Documentation

data Token

This type represents an extracted token returned by unexpectedToken in ErrorBuilder.

There is deliberately no analogue for EndOfInput because we guarantee that non-empty residual input is provided to token extraction.

Constructors

Raw
This is a token that is directly extracted from the residual input itself.

Fields

!String

the input extracted.

Named
This is a token that has been given a name, and is treated like a labelled item.

Fields

!String

the description of the token.

!Word

the amount of residual input this token ate.

type TokenExtractor = NonEmpty Char -> Word -> Bool -> Token

tillNextWhitespace :: Bool -> (Char -> Bool) -> TokenExtractor

singleChar :: TokenExtractor

matchParserDemand :: TokenExtractor