Text.Gigaparsec.Internal

Description

This module does not adhere to PVP, and can change at any time as required by the maintainers of the library. Use this functionality at your own risk.

Since: 0.1.0.0

Documentation

data State

Constructors

State

Fields

input :: !String

the input string, in future this may be generalised

consumed :: !Bool

has the parser consumed input since the last relevant handler?

line :: !Int

the current line number (incremented by n)

col :: !Int

the current column number (have to settle on a tab handling scheme)

newtype Parsec a

Constructors

Parsec

Fields

unParsec :: forall r. State -> (a -> State -> r) -> (State -> r) -> r

emptyState :: String -> State

_branch :: Parsec (Either a b) -> Parsec (a -> c) -> Parsec (b -> c) -> Parsec c

This is an internal implementation of branch, which is more efficient than the Selective default branch. We should be using this internally, and it can be dropped if https://github.com/snowleopard/selective/issues/74 is implemented.