Text.Gigaparsec.Patterns

Description

This module is currently experimental, and may have bugs depending on the version of Haskell, or the extensions enabled. Please report any issues to the maintainers.

Since: 0.2.2.0

Documentation

deriveLiftedConstructors :: String -> [Name] -> Q [Dec]

deriveDeferredConstructors :: String -> [Name] -> Q [Dec]