Text.Gigaparsec.Token.Patterns

Description

This module is currently experimental, and may have bugs depending on the version of Haskell, or the extensions enabled. Please report any issues to the maintainers.

Since: 0.2.2.0

Documentation

overloadedStrings

Arguments

:: Q Exp
the quoted Lexer

-> Q [Dec]
a synthesised IsString instance.

When given a quoted reference to a Lexer, for example [|lexer|], this function will synthesise an IsString instance that will allow string literals to serve as Parsec (). These literals will parse symbols in the language associated with the lexer, followed by consuming valid whitespace.

Since: 0.2.2.0